Move callables into their own top-level directory.
