+++ edu_edfi_airflow/dags/edfi_resource_dag.py
@@ -11,8 +11,7 @@ from ea_airflow_util import EACustomDAG
 from ea_airflow_util import slack_callbacks, update_variable
 from edfi_api_client import camel_to_snake
 
-from edu_edfi_airflow.dags.callables import change_version
-from edu_edfi_airflow.dags.dag_util import airflow_util
+from edu_edfi_airflow.callables import change_version, airflow_util
 from edu_edfi_airflow.providers.edfi.transfers.edfi_to_s3 import EdFiToS3Operator
 from edu_edfi_airflow.providers.snowflake.transfers.s3_to_snowflake import S3ToSnowflakeOperator
