+++ edu_edfi_airflow/providers/earthbeam/operators.py
@@ -6,7 +6,7 @@ from typing import Iterable, Optional, Union
 from airflow.models import Connection
 from airflow.operators.bash import BashOperator
 
-from edu_edfi_airflow.dags.dag_util import airflow_util
+from edu_edfi_airflow.callables import airflow_util
 
 
 class EarthmoverOperator(BashOperator):
